+++ apps/src/features/manager-selection/index.ts
@@ -1,0 +1,1 @@
+export * from './lib';
